event_site/requirement.txt
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.1
django-crispy-forms==1.7.2
ecdsa==0.13.2
future==0.17.1
html5lib==1.0.1
idna==2.8
oauthlib==3.0.1
Pillow==6.0.0
pyasn1==0.4.5
pycryptodomex==3.8.1
pyjwkest==1.4.0
PyJWT==1.7.1
PyPDF2==1.26.0
python-gnupg==0.4.4
python-jose==3.0.1
python3-openid==3.1.0
pytz==2019.1
reportlab==3.5.23
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-backend-epita==0.0.4
social-auth-core==3.1.0
sphinx
sqlparse==0.3.0
urllib3==1.24.3
webencodings==0.5.1
xhtml2pdf==0.2.3
